Refactor Cardano transaction hash calculation
This commit replaces the manual transaction hash calculation in `executeCardanoTx` with a new utility function, `getCardanoTxHash`, improving code readability and maintainability. The removed code included detailed logic for parsing CBOR and calculating the transaction ID, which is now encapsulated in the new function.
